[BUGFIX beta] Fix {{each}} with `itemViewClass` and {{else}}.
